backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic[email]
python-dotenv==1.0.1
faker==24.14.0
httpx==0.27.0          # for async HTTP calls (Google Maps, etc.)
geopy==2.4.1           # haversine distance calculation
alembic==1.13.1        # DB migrations
psycopg2-binary==2.9.9 # PostgreSQL driver
python-multipart==0.0.9
slowapi==0.1.9         # rate limiting
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
